+++ packages/derivatives-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.62](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.61...@injectivelabs/derivatives-consumer@0.0.62) (2021-10-15)
+
+
+### Features
+
+* portfolio value ([0f85db1](https://github.com/InjectiveLabs/injective-ts/commit/0f85db13146cad0d1c4cb991e2190062e6156b71))
+
+
+
+
+
 ## [0.0.61](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.60...@injectivelabs/derivatives-consumer@0.0.61) (2021-10-14)
 
 **Note:** Version bump only for package @injectivelabs/derivatives-consumer